Move files around, fix includes
