+++ examples/fabric-bridge-app/fabric-bridge-common/BridgedAdministratorCommissioning.h
