+++ examples/fabric-bridge-app/fabric-bridge-common/BridgedDevice.h
