+++ examples/fabric-bridge-app/fabric-bridge-common/BridgedDeviceBasicInformationImpl.cpp
