+++ examples/fabric-bridge-app/fabric-bridge-common/BridgedDeviceManager.h
